Implement q_insert_head & q_insert_tail
First, check if the head is true. Then, use malloc() to get the memory address for the new element_t, also check if the malloc success. Assign string s to the value. In the end, use list_all or ist_add or list_add_tail to implement the insert_head or tail.